address comments, remove duplicated code, update unit tests
